display criteria filter div unconditionally
With addition of text attribute filters, the "add new filter" link should
be displayed even if the type has no reference attributes (all types are
assumed to have at least one text attribute, as the prefLabel is generated
automatically upon creation).

It is also possible that a type has no reference attributes, so the separator
in the dropdown list should instead be displayed only if reference  attributes
exist.