fix text attribute node lookup problem
Issue: The text attribute filter does not return all existing values because
the grouping of values is done after the query (default query with max 25
nodes returned). In some cases the nodes returned may all have the same value
for the text attribute. Looking up all nodes is also not desirable as
this slows down performance too much.

Solution: Iterative node lookup loop where each consecutive search query
excludes values already in the results. The loop terminates when no more
nodes are returned or when there are 25 results (beyond that it is reasonable
to expect that the user refines the search terms).